requirements.txt
aiogram
python-dotenv
dotenv
psycopg2-binary
aiogram-dialog
apscheduler
pandas